add common vector DB base class
